test(proto): split server default config coverage
